move main page into separate file
